requirements/common.txt
celery==3.1.20
Django==1.9.1
django-celery==3.1.17
django-gravatar2==1.4.0
django-role-permissions==1.0
djangorestframework==3.3.2
redis==2.10.5
selenium==2.52.0
django-configurations==1.0
django_compressor==2.0